IntelliHire_AI/requirements.txt
# Core
python-dotenv>=1.0.0

# Resume parsing
PyMuPDF>=1.23.0
python-docx>=1.1.0

# LLM (Gemini)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0

# Matching
sentence-transformers>=2.6.0
scikit-learn>=1.4.0
numpy>=1.26.0


deepgram-sdk 
sounddevice 
websockets
python-dotenv
numpy
aiohttp
aiortc
asyncio
scipy
fastapi 
uvicorn 
av
aiohttp_cors
pydub
# caching resumes
redis
python-multipart
